backend/requirements/base.txt
Django==4.2.11
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-environ==0.11.2
drf-spectacular==0.27.2
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-redis==5.4.0
celery==5.3.6
redis==5.0.3
psycopg[binary]==3.1.18
whitenoise==6.6.0
anthropic==0.25.1
textblob==0.18.0
Pillow==10.3.0
gunicorn==21.2.0
